+++ application/ticket-app-api/build.gradle
@@ -3,10 +3,6 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 
-	//spring-security
-	implementation 'org.springframework.boot:spring-boot-starter-security'
-	testImplementation 'org.springframework.security:spring-security-test'
-
 	//spring-data
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
