+++ application/ticket-app-api/src/main/java/com/uket/app/ticket/api/config/BaseConfig.java
@@ -6,4 +6,5 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 @Configuration
 @EnableJpaAuditing
 public class BaseConfig {
+
 }
