+++ application/ticket-app-api/src/main/java/com/uket/app/ticket/api/controller/DevApi.java
@@ -1,8 +1,9 @@
 package com.uket.app.ticket.api.controller;
 
-import com.uket.app.ticket.api.config.userid.LoginUserId;
+import com.uket.domain.auth.config.userid.LoginUserId;
 import com.uket.app.ticket.api.dto.response.TokenResponse;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,7 +18,9 @@ public interface DevApi {
     @Operation(summary = "테스트용 api", description = "테스트를 진행합니다.")
     @SecurityRequirement(name = "JWT")
     ResponseEntity<String> test(
-            @LoginUserId Long userId
+            @Parameter(hidden = true)
+            @LoginUserId
+            Long userId
     );
 
     @GetMapping("/token")
