--- application/ticket-app-api/src/main/java/com/uket/app/ticket/api/filter/JwtAccessDeniedHandler.java
@@ -1,36 +1,0 @@
-package com.uket.app.ticket.api.filter;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.uket.core.dto.response.ErrorResponse;
-import com.uket.core.exception.ErrorCode;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response,
-            AccessDeniedException accessDeniedException) throws IOException, ServletException {
-        writeErrorResponse(response, accessDeniedException);
-    }
-
-    private void writeErrorResponse(
-            HttpServletResponse response, AccessDeniedException accessDeniedException
-    ) throws IOException {
-        response.setCharacterEncoding(StandardCharsets.UTF_8.name());
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setStatus(HttpServletResponse.SC_FORBIDDEN);
-        response.getWriter().write(objectMapper.writeValueAsString(ErrorResponse.of(ErrorCode.AUTHORIZATION_FAILED)));
-    }
-}
