+++ domain/auth-domain/src/main/java/com/uket/domain/auth/config/userid/LoginUserArgumentResolver.java
@@ -1,6 +1,5 @@
-package com.uket.app.ticket.api.config.userid;
+package com.uket.domain.auth.config.userid;
 
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.MethodParameter;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
@@ -10,7 +9,6 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
 @Component
-@Slf4j
 public class LoginUserArgumentResolver implements HandlerMethodArgumentResolver {
 
     @Override
@@ -23,7 +21,6 @@ public class LoginUserArgumentResolver implements HandlerMethodArgumentResolver 
             NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
 
         String userId = SecurityContextHolder.getContext().getAuthentication().getName();
-        log.info("userid : {}",userId);
         return Long.parseLong(userId);
     }
 }
