+++ domain/auth-domain/src/main/java/com/uket/domain/auth/config/userid/LoginUserId.java
@@ -1,12 +1,10 @@
-package com.uket.app.ticket.api.config.userid;
+package com.uket.domain.auth.config.userid;
 
-import io.swagger.v3.oas.annotations.Hidden;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Hidden
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface LoginUserId {
