+++ domain/auth-domain/src/main/java/com/uket/domain/auth/dto/response/AuthToken.java
@@ -5,6 +5,7 @@ public record AuthToken(
         String refreshToken,
         Boolean isRegistered
 ) {
+
     public static AuthToken of(String accessToken, String refreshToken, Boolean isRegistered) {
         return new AuthToken(accessToken, refreshToken, isRegistered);
     }
