+++ domain/auth-domain/src/main/java/com/uket/domain/auth/dto/response/userinfo/OAuth2UserInfoResponse.java
@@ -1,8 +1,12 @@
 package com.uket.domain.auth.dto.response.userinfo;
 
 public interface OAuth2UserInfoResponse {
+
     String getProvider();
+
     String getProviderId();
+
     String getEmail();
+
     String getName();
 }
