+++ domain/auth-domain/src/main/java/com/uket/domain/auth/filter/JWTTokenAuthentication.java
@@ -1,4 +1,4 @@
-package com.uket.app.ticket.api.filter;
+package com.uket.domain.auth.filter;
 
 import com.uket.domain.user.dto.UserDto;
 import java.util.ArrayList;
